choar: refactor higlight feature
